Move resource.Info to its own file.
